+++ Coimbra/Services/IApplicationService.cs
@@ -1,7 +1,10 @@
-﻿namespace Coimbra
+﻿using UnityEngine.Scripting;
+
+namespace Coimbra
 {
     /// <summary>
     /// Responsible for the application lifetime cycle, meant to be used to fire Unity's callbacks events.
     /// </summary>
+    [RequireImplementors]
     public interface IApplicationService : IService { }
 }